requirements.txt
faker>=37.0.0
playwright>=1.40.0
pyautogui>=0.9.54
pyperclip>=1.8.2
requests>=2.31.0
pytesseract>=0.3.10
Pillow>=10.0.0
